Changed runtime to 12 hours for testing. Changed CSV output module to output float values.

+++ src/ca/brood/softlogger/dataoutput/CsvOutputModule.java
@@ -136,7 +136,7 @@ public class CsvOutputModule extends AbstractOutputModule {
 			if (r.isNull())
 				values.add("NULL");
 			else {
-				values.add(""+r.getInteger());
+				values.add(""+r.getFloat());
 				atLeastOneGoodValue = true;
 			}
 		}
